Resolve "Finir l'implémentation des tests des controllers"

Closes #37

See merge request Sacane/ugeoverflow!45

+++ RevevueApp/src/main/java/fr/pentagon/ugeoverflow/controllers/dtos/responses/VoteDTO.java
@@ -1,6 +1,6 @@
 package fr.pentagon.ugeoverflow.controllers.dtos.responses;
 
-public record VoteDTO(int upVotes, int downVotes) {
+public record VoteDTO(long upVotes, long downVotes) {
 
     public VoteDTO {
         if(upVotes < 0 || downVotes < 0)
